Move dialog classes to their own directory
